crews/research_basic/research.py
from crewai_tools import BaseTool
import requests
from bs4 import BeautifulSoup
import arxiv
from dotenv import load_dotenv

load_dotenv()

class ComprehensiveResearchTool(BaseTool):
    name: str = "ComprehensiveResearchTool"
    description: str = "A tool for browsing the internet, scraping websites, and searching arXiv."

    def _run(self, argument: dict) -> str:
        # Implementation will call the appropriate method based on the argument
        if argument['action'] == 'browse':
            return self.browse_internet(argument['query'])
        elif argument['action'] == 'scrape':
            return self.scrape_website(argument['url'])
        elif argument['action'] == 'search_arxiv':
            return self.search_arxiv(argument['query'])
        else:
            return "Invalid action specified"

    def browse_internet(self, query: str) -> str:
        # Placeholder for browsing implementation
        # Use requests and BeautifulSoup or similar tools
        return "Mock browsing result for query: " + query
    
    def scrape_website(self, url: str) -> str:
        # Placeholder for scraping implementation
        # Use BeautifulSoup for scraping the given URL
        return "Mock scraping result for URL: " + url
    
    def search_arxiv(self, query: str) -> str:
        # Use the arxiv package to search for publications
        search = arxiv.query(query=query, max_results=5)
        titles = [article['title'] for article in search]
        return "Found articles: " + ", ".join(titles)



from crewai import Agent, Crew, Task, Process
from crewai_tools import tool  # Placeholder for actual tool import

# Initialize the comprehensive research tool
research_tool = ComprehensiveResearchTool()

# Agents
internet_researcher = Agent(
    role='Internet Researcher',
    goal='Gather broad insights and trends on {topic}',
    tools=[research_tool],
    memory=True,
    verbose=True
)

publication_researcher = Agent(
    role='Publication Researcher',
    goal='Integrate major research publications insights on {topic}',
    tools=[research_tool],
    memory=True,
    verbose=True
)

social_media_analyst = Agent(
    role='Social Media Analyst',
    goal='Research social media trends on {topic}',
    tools=[research_tool],
    memory=True,
    verbose=True
)

# Logging interactions - This will be part of each agent's operation, ensuring all interactions are captured.


# Tasks - Placeholder tasks, real tasks should be defined based on actual research needs
research_task = Task(
    description="Comprehensive research on {topic} for CFO considerations in 2024.",
    expected_output="A detailed report on business considerations for CFOs on {topic}.",
    tools=[research_tool],
    agent=internet_researcher,  # This should be assigned dynamically based on task requirements
)

# Crew formation
research_crew = Crew(
    agents=[internet_researcher, publication_researcher, social_media_analyst],
    tasks=[research_task],
    process=Process.sequential,  # Using sequential process for task execution
)

# To kick off the crew with a specific topic:
result = research_crew.kickoff(inputs={'topic': 'Sustainable Energy Initiatives'})
print(result)